# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Work Manager API
  version: 1.0.0
  description: |
    The two token-authenticated REST surfaces of this application. Everything here is **verified against the
    running code on every test run** (`tests/Functional/Api/OpenApiContractTest.php`): each operation is called
    for real and its status codes, headers and JSON bodies are validated against the schemas below, and
    `OpenApiSpecTest` fails if a route exists that is not documented (or the reverse). If this file and the code
    disagree, the build is red — fix whichever one is wrong.

    ## Authentication — both APIs

    Both surfaces use the same **personal access tokens** (`auth-pat-bundle`): create one at `/account/tokens` and
    send it as `Authorization: Bearer <token>`. The plaintext is shown once; only its hash is stored. A revoked,
    expired or unknown token, or one whose owner is deactivated, is answered `401`. What a token may do follows its
    owner's roles (ADR-068): the Admin API needs an account holding an admin role (else `403`); some endpoints need
    tech support. Without `auth-pat-bundle` no token can authenticate, so both APIs refuse every call (`401`).

    ## Admin API — `/admin-api/*`

    Account and instance management, for accounts holding `ROLE_ADMIN` or above. An admin manages regular users;
    a super admin also manages admins; tech-support accounts are visible only to tech support.

...

    ## User API — `/api/*`

...

...

...
servers:
...
paths:
  /admin-api/users:
    get:
      operationId: listUsers
      tags: [Admin users]
      summary: List users
      description: Paginated and filterable. `per_page` outside 1–100 is clamped, not rejected; an unrecognised `status` is ignored.
      security:
        - adminBearer: []
      parameters:
        - name: page
          in: query
          description: 1-based page number, clamped to 1–1000000 (values below 1 become 1, above 1000000 become 1000000).
          schema: { type: integer, minimum: 1, maximum: 1000000, default: 1 }
          example: 1
        - name: per_page
          in: query
          description: Page size, clamped to 1–100.
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
          example: 10
        - name: email
          in: query
          description: Substring filter on the email address.
          schema: { type: string }
          example: example.com
        - name: status
          in: query
          description: Only users with this status.
          schema: { $ref: '#/components/schemas/AccountStatus' }
          example: active
      responses:
        '200':
          description: A page of users.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserList' }
              example:
                data:
                  - id: 7
                    email: jane.doe@example.com
                    name: Jane Doe
                    status: active
                    roles: [ROLE_USER]
                    created_at: '2026-09-30T21:20:14+00:00'
                meta: { total: 1, page: 1, per_page: 10, total_pages: 1 }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      operationId: createUser
      tags: [Admin users]
      summary: Create a user
      description: |
        Validates exactly like the admin UI (shared service): email format and uniqueness, required name,
        password policy, known role/status. `role` is fixed to `ROLE_USER` for users.
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
            example:
              email: jane.doe@example.com
              name: Jane Doe
              password: CorrectHorse-Battery9Staple!
              role: ROLE_USER
              status: active
      responses:
        '201':
          description: The created user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
              example:
                id: 7
                email: jane.doe@example.com
                name: Jane Doe
                status: active
                roles: [ROLE_USER]
                created_at: '2026-09-30T21:20:14+00:00'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /admin-api/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      operationId: getUser
      tags: [Admin users]
      summary: Get a user
      security:
        - adminBearer: []
      responses:
        '200':
          description: The user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UserNotFound' }

    patch:
      operationId: updateUser
      tags: [Admin users]
      summary: Update a user
      description: |
        PATCH semantics — only the keys you send are validated and applied (`email`, `name`, `role`, `status`).
        Setting `status` to `inactive` cuts access immediately (sessions, tokens and recovery links are revoked).
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
            example:
              name: Jane Q. Doe
      responses:
        '200':
          description: The updated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UserNotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }

    delete:
      operationId: deleteUser
      tags: [Admin users]
      summary: Delete (deactivate) a user
      description: Soft delete — sets `status` to `inactive` and revokes outstanding recovery tokens. The record is kept.
      security:
        - adminBearer: []
      responses:
        '204': { description: Deleted (deactivated). No body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UserNotFound' }
...
  /admin-api/users/{id}/password-reset:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      operationId: sendUserPasswordReset
      tags: [Admin users]
      summary: Email a password-reset link
      security:
        - adminBearer: []
      responses:
        '200':
          description: The reset link was queued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOk' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UserNotFound' }
...
  /admin-api/users/{id}/unlock:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      operationId: unlockUser
      tags: [Admin users]
      summary: Clear an account lockout
      description: '**Bundle:** `auth-security-bundle`. A no-op when the account is not locked.'
      x-bundle: auth-security-bundle
      security:
        - adminBearer: []
      responses:
        '200':
          description: The account is unlocked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOk' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UserNotFound' }
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      description: A personal access token (created at `/account/tokens`) whose owner holds an admin role.
    userBearer:
      type: http
      scheme: bearer
      description: A user's personal access token, created at `/account/tokens`.
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: User id.
      schema: { type: integer, minimum: 1 }
      example: 7
  responses:
    Unauthorized:
      description: The bearer token is missing, invalid, revoked, expired, or belongs to an inactive account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Missing or invalid Authorization header. }
    UserNotFound:
      description: No user with that id.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: User not found. }
    ValidationFailed:
      description: One or more fields failed validation. Keys are field names.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrors' }
          example:
            errors:
              email: This email address is already registered.
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string, minLength: 1 }
    ValidationErrors:
      type: object
      additionalProperties: false
      required: [errors]
      properties:
        errors:
          type: object
          minProperties: 1
          additionalProperties: { type: string, minLength: 1 }
    StatusOk:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { const: ok }
    AccountStatus:
      type: string
      enum: [active, inactive]
    User:
      type: object
      additionalProperties: false
      required: [id, email, name, status, roles, created_at]
      properties:
        id: { type: integer, minimum: 1 }
        email: { type: string, format: email }
        name: { type: string }
        status: { $ref: '#/components/schemas/AccountStatus' }
        roles:
          type: array
          minItems: 1
          items: { type: string, enum: [ROLE_USER] }
        created_at: { type: string, format: date-time }
    UserList:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/User' }
        meta:
          type: object
          additionalProperties: false
          required: [total, page, per_page, total_pages]
          properties:
            total: { type: integer, minimum: 0 }
            page: { type: integer, minimum: 1 }
            per_page: { type: integer, minimum: 1, maximum: 100 }
            total_pages: { type: integer, minimum: 1 }
...
    CreateUserRequest:
      type: object
      required: [email, name, password]
      properties:
        email: { type: string, format: email }
        name: { type: string, minLength: 1 }
        password: { type: string, description: Must satisfy the configured password policy. }
        role: { type: string, enum: [ROLE_USER], default: ROLE_USER }
        status: { $ref: '#/components/schemas/AccountStatus' }
    UpdateUserRequest:
      type: object
      minProperties: 1
      properties:
        email: { type: string, format: email }
        name: { type: string, minLength: 1 }
        role: { type: string, enum: [ROLE_USER] }
        status: { $ref: '#/components/schemas/AccountStatus' }
...